dad3d/requirements.txt
wheel==0.37.1
hydra-core==1.1.0
coloredlogs

numpy==1.22.0
scikit-learn>=0.21.3
chumpy==0.70

#torch==1.9.0
#torchvision==0.10.0
#pytorch-lightning==1.6.0
torch_optimizer==0.1.0
pytorchcv==0.0.65
pytorch-toolbelt==0.5.0
timm==0.4.5
albumentations==1.0.0
smplx==0.1.26
fire
